Restructure emacs_init.el to remove cruft, implement explicit directory tracking
Since starship does so much funky stuff with the command prompt we
need to handle the prompt parsing for emacs shell mode ourselves. On
each directory tracking call split the lines, get the last line,
extract the ANSI escape codes and parse out the current
directory. Current directory has to be at the start of the last line.

ANSI escape codes are cleaned but emojis are not. Be wary of any put
into the last line by starship.